requirements.txt
msal
requests
openpyxl
python-dateutil
